doc(reverse-integer): add the description and time/space complex as comments
